+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## NEXT
+
+* `config`: Allow to configure a delay between commands
+
 ## 1.9.0 (2023-08-08)
 
 * `command`: A new field `msg.responseStatus` has been added to the response